requirements.txt
tensorflow
bokeh
numpy
pandas
streamlit
joblib
pathlib
nltk
keras-preprocessing
imagify
google-generativeai